Make sst/vectorAdd runnable

+++ util/job_launching/run_simulations.py
@@ -49,6 +49,7 @@ class ConfigurationSpec:
         for dir_bench in benchmarks:
             exec_dir, data_dir, benchmark, self.command_line_args_list = dir_bench
             full_exec_dir = os.path.join( this_directory, exec_dir )
+            full_sst_exec = os.path.join( this_directory, exec_dir, benchmark, benchmark)
             full_data_dir = os.path.join( this_directory, data_dir, benchmark.replace('/','_') )
 
             self.benchmark_args_subdirs = {}
@@ -59,8 +60,8 @@ class ConfigurationSpec:
             for args in self.command_line_args_list:
                 appargs_run_subdir = os.path.join( benchmark.replace('/','_'),
                                 self.benchmark_args_subdirs[args] )
-                this_run_dir = os.path.join( run_directory, appargs_run_subdir, self.run_subdir )
-                self.setup_run_directory(full_data_dir, this_run_dir, data_dir, appargs_run_subdir)
+                this_run_dir = os.path.join( run_directory, appargs_run_subdir, self.run_subdir)
+                self.setup_run_directory(full_data_dir, this_run_dir, data_dir, appargs_run_subdir, full_sst_exec)
 
                 self.text_replace_torque_sim(full_data_dir,this_run_dir,benchmark,cuda_version, args, libdir, full_exec_dir,build_handle)
                 if not options.run_sst:
@@ -120,7 +121,7 @@ class ConfigurationSpec:
     # Internal utility methods
     #########################################################################################
     # copies and links the necessary files to the run directory
-    def setup_run_directory(self, full_data_dir, this_run_dir, data_dir, appargs_subdir):
+    def setup_run_directory(self, full_data_dir, this_run_dir, data_dir, appargs_subdir, full_sst_exec):
         if not os.path.isdir(this_run_dir):
             os.makedirs(this_run_dir)
 
@@ -143,8 +144,10 @@ class ConfigurationSpec:
             shutil.copyfile(file_to_cp,new_file)
 
         # link the executable directory
-        #if options.run_sst:
-
+        if options.run_sst:
+            if os.path.lexists(os.path.join( this_run_dir, os.path.                         basename(full_sst_exec) ) ):
+                os.remove(os.path.join( this_run_dir, os.path.                         basename(full_sst_exec) ) )
+            os.symlink(full_sst_exec, os.path.join( this_run_dir, os.path.basename(full_sst_exec) ) )
 
         # link the data directory
         benchmark_data_dir = os.path.join(full_data_dir, "data")
@@ -161,12 +164,13 @@ class ConfigurationSpec:
                     os.remove(os.path.join(this_run_dir, "traces"))
                 os.symlink(benchmark_trace_dir, os.path.join(this_run_dir,"traces"))
 
+        if not options.run_sst:
+            all_data_link = os.path.join(this_run_dir,"data_dirs")
+            if os.path.lexists(all_data_link):
+                os.remove(all_data_link)
+            if os.path.exists(os.path.join(this_directory, data_dir)):
+                os.symlink(os.path.join(this_directory, data_dir), all_data_link)
 
-        all_data_link = os.path.join(this_run_dir,"data_dirs")
-        if os.path.lexists(all_data_link):
-            os.remove(all_data_link)
-        if os.path.exists(os.path.join(this_directory, data_dir)):
-            os.symlink(os.path.join(this_directory, data_dir), all_data_link)
     # replaces all the "REAPLCE_*" strings in the torque.sim file
     def text_replace_torque_sim( self,full_run_dir,this_run_dir,benchmark, cuda_version, command_line_args,
                                  libpath, exec_dir, gpgpusim_build_handle ):
